+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added enhancements to the start/stop service to select all services or specific ones.
+
 ## [0.2.44] - 2025-01-22
 
 ### Added