+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added enhancements to the start/stop service to select all services or specific ones.
 - Added enhancement to update_image service to also update the sidecar container image.
+- Added support for service rollback with a specific revision.
 
 ## [0.2.44] - 2025-01-22